requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
PyJWT==2.8.0
bcrypt==3.2.2
passlib[bcrypt]==1.7.4
langchain==0.1.1
langchain-groq==0.0.1
python-multipart==0.0.6
aiofiles==23.2.1
PyPDF2==3.0.1
